requirements.txt
Package                       Version
----------------------------- ---------
amqp                          5.1.1
asgiref                       3.4.1
attrs                         22.2.0
billiard                      3.6.4.0
cached-property               1.5.2
celery                        5.1.2
certifi                       2022.9.24
cffi                          1.15.1
charset-normalizer            2.0.12
click                         7.1.2
click-didyoumean              0.3.0
click-plugins                 1.1.1
click-repl                    0.2.0
coreapi                       2.3.3
coreschema                    0.0.4
cryptography                  38.0.4
defusedxml                    0.7.1
Django                        3.2.16
django-celery-results         2.2.0
django-cors-headers           3.10.1
django-templated-mail         1.1.1
djangorestframework           3.14.0
djangorestframework-simplejwt 4.4.0
djoser                        2.1.0
drf-yasg                      1.21.4
flake8                        5.0.4
idna                          3.4
importlib-metadata            1.7.0
inflection                    0.5.1
iniconfig                     1.1.1
itypes                        1.2.0
Jinja2                        3.0.3
kombu                         5.1.0
MarkupSafe                    2.0.1
mccabe                        0.7.0
oauthlib                      3.2.2
packaging                     21.3
pip                           21.3.1
pkg_resources                 0.0.0
pluggy                        1.0.0
prompt-toolkit                3.0.36
py                            1.11.0
pycodestyle                   2.9.1
pycparser                     2.21
pyflakes                      2.5.0
PyJWT                         2.4.0
pyparsing                     3.0.9
pytest                        7.0.1
pytest-django                 4.5.2
python-dotenv                 0.20.0
python3-openid                3.2.0
pytz                          2022.6
requests                      2.27.1
requests-oauthlib             1.3.1
ruamel.yaml                   0.17.21
ruamel.yaml.clib              0.2.7
setuptools                    39.0.1
six                           1.16.0
social-auth-app-django        4.0.0
social-auth-core              4.3.0
sqlparse                      0.4.3
tomli                         1.2.3
typing_extensions             4.1.1
uritemplate                   4.1.1
urllib3                       1.26.13
vine                          5.0.0
wcwidth                       0.2.5
zipp                          3.6.0
